Create config dir in root directory and rename src/config to src/constants.
