+++ config/webpack.prod.js
