+++ package.json
@@ -5,9 +5,9 @@
   "main": "dist/index.html",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --config webpack.dev.js",
+    "start": "webpack-dev-server --config config/webpack.dev.js",
     "linter": "eslint **/*.js **/*.jsx",
-    "build": "webpack --config webpack.prod.js"
+    "build": "webpack --config config/webpack.prod.js"
   },
   "author": "Dmytro Paniontko",
   "license": "ISC",
